+++ src/main/java/com/gdu/wacdo/services/AffectationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gdu.wacdo.models.Affectation;
 
-public interface AffectationRepository extends JpaRepository<Affectation, Integer>{
+public interface AffectationRepository extends JpaRepository<Affectation, Long>{
 
 }
