+++ src/main/java/com/gdu/wacdo/services/RestaurantsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gdu.wacdo.models.Restaurant;
 
-public interface RestaurantsRepository extends JpaRepository<Restaurant,  Integer> {
+public interface RestaurantsRepository extends JpaRepository<Restaurant,  Long> {
 
 }
